update values conversion

+++ DataControl/GPSData.cs
@@ -12,18 +12,18 @@ namespace MavBoia.DataControl
         private double latitude; /*< [degE7] */
         private double longitude; /*< [degE7] */
         private float speed; /*< [cm/s] */
-        private float course; /*< [deg] Direction of movement(not heading)*/
-        private float heading; /*< [deg] Direction of the bow (source not from GPS, but from IMU)*/
-        private float satellites_visible; /*<  */
-        private int hdop; /*<  Horizontal dilution of precision. Represents quality of GPS constellation. Values close to 1 are best. Above 5 is poor*/
+        private int course; /*< [deg] Direction of movement(not heading)*/
+        private int heading; /*< [deg] Direction of the bow (source not from GPS, but from IMU)*/
+        private int satellites_visible; /*<  */
+        private float hdop; /*<  Horizontal dilution of precision. Represents quality of GPS constellation. Values close to 1 are best. Above 5 is poor*/
 
         public double Latitude { get => latitude; set => latitude = value; }
         public double Longitude { get => longitude; set => longitude = value; }
         public float Speed { get => speed; set => speed = value; }
-        public float Course { get => course; set => course = value; }
-        public float Heading { get => heading; set => heading = value; }
-        public float Satellites_visible { get => satellites_visible; set => satellites_visible = value; }
-        public int Hdop { get => hdop; set => hdop = value; }
+        public int Course { get => course; set => course = value; }
+        public int Heading { get => heading; set => heading = value; }
+        public int Satellites_visible { get => satellites_visible; set => satellites_visible = value; }
+        public float Hdop { get => hdop; set => hdop = value; }
 
         public GPSData() { }
         public GPSData(MAVLink.mavlink_gps_t mavGps)
